+++ S2ZonaConfigurator/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## [0.0.5]
 - Add a diff report generation feature
-- 
+- Add an ability specify the file path only once in JSON mods if the file is used in multiple actions
+- Fix: Some issues when last actions may not be applied in mods where multiple files are used
+
 ## [0.0.4]
 - Speed up the mods processing
 - Fix: Modify multiple values method may produce the wrong indentation in a modified line by duplicating a part of the key name